ft(user): syn front end with backend
